vider-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
torch
transformers
accelerate
sentencepiece
python-dotenv
pydantic
